fix/move detail component
